refactor(LottoGame) : 서비스 패키지로 이동
